Always write thumbnails to canvas
When an image is smaller than the target thumbnail size, it was not drawn
to a canvas. The `<img/>` object was returned unchanged. This caused
issues further on in the code because it tried to convert the resized
canvas to a Blob.

This patch also draws small images to a canvas inside `protect()`. This
way we always end up with a canvas element, fixing the above issue and
also making sure that GIFs don't play and backgrounds are not
transparent (for consistency).